pyproject.toml
[tool.poetry]
name = "continuum"
version = "0.1.0"
description = ""
authors = ["kevin hill <kivo360@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.5.0"
torchvision = "^0.6.0"
gym = "^0.17.2"
matplotlib = "^3.2.1"
gpytorch = "^1.1.1"
tqdm = "^4.46.1"
loguru = "^0.5.1"
creme = "^0.6.1"
numpy = "^1.19.0"
ta = "^0.5.25"
toolz = "^0.10.0"
sklearn = "^0.0"
pydantic = "^1.6.1"
devtools = "^0.6"
maya = "^0.6.1"
einops = "^0.3.0"
addict = "^2.3.0"
jsonschema = "^3.2.0"
python-digitalocean = "^1.15.0"
tensorboard = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.5.3"
mypy = "^0.781"
flake8 = "^3.8.3"
autopep8 = "^1.5.3"
yapf = "^0.30.0"
jupyterlab = "^2.2.4"
jupyter-lsp = "^0.9.1"
python-language-server = { extras = ["all"], version = "^0.34.1" }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
